requirements.txt
click==5.1
nose==1.3.4
numpy==1.9.2
pbr==0.10.0
Pillow==2.9.0
six==1.8.0
stevedore==1.1.0
virtualenv==1.11.6
virtualenv-clone==0.2.5
